+++ packages/react-jss/src/createUseStyles.test.js
@@ -1,5 +1,4 @@
 /* eslint-disable react/prop-types */
-
 import createUseStyles from './createUseStyles'
 import createBasicTests from '../test-utils/createBasicTests'